Import Aliasing: Use if basename doesn't match
This adds a note to the import aliasing section that import aliasing
must be used if the base name doesn't match the package name.

This is placed before the more general recommendation of avoiding import
aliasing otherwise because it's more important to use named imports when
the basename doesn't match.