Clarify guidance for option functions. (#11)
Consistent with our guidance to avoid providing public functions that
accept private types and to give library authors full latitude to change
the underlying implementation of functional options, we now suggest
using a public Option interface with a private method.

We have also learned not to mix the words option and configuration in
the same context. An option is a function parameter to override
behavior. Configuration is parsed from files and frequently maps to
options.