Reduce Scope: Fix FD leak
As pointed out in #19, the second Good example for "Reduce Scope of
Variables" leaks an open file descriptor if `io.WriteString` fails.

Further, both sets of examples are manually closing files instead of
using `defer` as pointed out elsewhere in the guide.

As these examples just need a pair of functions, one returning just an
error, another returning a value and an error, I've switched them to
using `ioutil.ReadFile`, `ioutil.WriteFile`, and a plausible
`Config.Decode([]byte) error` method on a `cfg` variable assumed to be
in-scope.

Resolves #19