Functional Options: Don't recommend closures
This changes the Functional Options example to not recommend closures,
instead favoring full implementations of the interface for each option.

To help demonstrate that options don't have to be new types, I added a
logger option to the example implemented around a `struct`. This made
the examples too long so I dropped the timeout option: I wanted to keep
cache around to demonstrate readability of naked booleans versus
`WithCache(false)`.

I renamed the function to Open to further reduce horizontal space
consumed by the examples.