Guidelines: Verify Interface Compliance
This adds advice on verifying interface compliance with
`var _ Interface = (*Foo)(nil)`.

As part of this guideline, we have to cover that if the receiver is on
the value type, the right side should be the zero value. Because of
that, the guideline is placed right before the section where we talk
about receiver types for interfaces.

Resovles #25