style: Lint with markdownlint
markdownlint is a static analysis tool with
a library of rules to enforce standards and
consistency for Markdown files. [^1]

This commit removes redundant spaces,
unused references, indent mistakes, etc.
from this document according to
the report from `markdownlint`.

Note that I disable `MD033` (no-
inline-html) as it is necessary for
side-by-side code samples.

[^1]: https://github.com/DavidAnson/markdownlint

Signed-off-by: pan93412 <pan93412@gmail.com>