Add recommendation on using table tests for single case tests.
The exact wording is TBD - this is to start a discussion.

My personal preference is to use table-test all the time, and would obviously 
would like it to be recommended.

However, the primary goal is to avoid arguing in the code reviews. I am okay
with the other version too (no table tests for single case tests), as long as we
call something out here.